+++ tests/test_all.py
@@ -25,6 +25,7 @@ import test_festa
 import test_bghost
 import test_forbidden
 import test_hiyari
+import test_hitogomi
 
 
 def suite():
@@ -48,6 +49,7 @@ def suite():
         unittest.makeSuite(test_disliked.StoryTest),
         unittest.makeSuite(test_festa.StoryTest),
         unittest.makeSuite(test_hiyari.StoryTest),
+        unittest.makeSuite(test_hitogomi.StoryTest),
         # next fantasy
         unittest.makeSuite(test_lv1yusha.StoryTest),
         # writing cheer