+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Project: Forbidden love
 - Project: Blue ghost
 - Project: Night festa
+- Project: Hitogomi
 
 ## [0.6.0] - 2019-08-01
 ### Added